Add review save-to-storage functionality.
This commit implements several changes in the Locations and IndexedDB
components to make saving reviews to stored IndexedDB locations easy.
The database now has an index on the `placeId` property. It also
has a clearer difference between its `query` and `filter` methods.
All the new functionality is wrapped up in the Locations component.