Change application menu state architecture.
This commit addresses a bug where clicking on a search result wouldn't remove
the search menu and site guard.

The change drops the old `pc-toggle` component in favor of an application-
level controller with properties. These properties map to class names on the
controller's root element; triggering property changes via the newly exposed
`setMenuState()` method changes classes on the element.

A new `pc-focus` directive is introduced to focus on the search input when
a visitor clicks the 'search' icon. This removes close coupling with the DOM and paves the way for removing jQuery as a dependency.